Update presenter based on song selection
Also allow selecting multiple songs.

Song planner selection should be loaded into the presenter as well as
into the presentation if it is active.

Every time a new song is added, the presenter and presentation is
reloaded and any current displayed song is no longer displayed.

This could be enhanced in the future as needed.