Blank screen when clicking top line in presenter
Upon clicking the title of a song or a line in the song in the
presenter, it is scrolled to the top in the presentation.

Upon clicking the same line again, it will now blank the
presentation (unslect the current song) while keeping the red
live frame in the same place of the presenter.

Closes #2