Swap in vw, aspect ratio scaling
Use the scaling in the presentation screen to maintain a specified
aspect ratio. Use vw units for font size in presenter and presentation.

Seeing an issue with the live frame not matching what is displayed in
the presentation. Needs further debugged.

Also need to compare performance. Scrolling seems maybe not as smooth.
Could be using scroll is smoother than negative margin-top.