Set presenter font size based on window width
Fix a bug where the presenter font size is not consistent. And with
varying window widths, lines are not wrapping consistently. And the live
frame is continually off, sometimes by a little, sometimes by a lot.

Resolve all of this by setting the font size of the presenter. The vw
units do not adjust for whether there is a scroll bar. So we need to
reduce the font size by the amount the viewport is wider than the
presenter box.