Fullscreen button even when song line selected
When the presentation window is opened, we want the Make Presentation
Fullscreen button to appear even if a song line is currently selected.
Previously, it would immediately hide the button. Instead keep the
button displayed and add a Dismiss button to hide the panel in case the
user doesn't want it to be fullscreen.