Add support for CCLI license info for songs
Be able to specify the license info in the songs with CCLI Song #: and
CCLI Copyrights: lines.

Only display songs with this info if the ccli.json file is correctly
populated.

Display copyright and CCLI license info on screen at the bottom of the
song.

Also include sorting of parsed songs, commented out.