Round trip changing of song selection
On changing the selection of the songs in the UI, restore back to the
previously selected songs. Then trigger the songs:change socket message
and wait for the response before then changing in the UI.

This prevents disconnected clients from attempting to change song
selection.

This is part of #10 of running presentation on separate device.