Initial calculation of wrap words for presenter
In presenter, calculate the wrap words for each clickable / wrapping
line that has some. This code will later be moved to an appropriate
place so there is a single authority on where the wraps should be and
all other clients will use white-space nowrap and insert <br>'s where
the wraps should be.